Move `ReusableSpecsSelector` to `spack.solver.reuse`
